change home page to server render component
